refactor(commonpb): move the contiguous network container to its own files
