+++ common/commonpb/meson.build
@@ -10,6 +10,7 @@ common_proto_files = [
     join_paths(common_proto_dir, 'v1', 'ipv6addr.proto'),
     join_paths(common_proto_dir, 'v1', 'iprange.proto'),
     join_paths(common_proto_dir, 'v1', 'ipnetwork.proto'),
+    join_paths(common_proto_dir, 'v1', 'contiguousnetwork.proto'),
     join_paths(common_proto_dir, 'v1', 'ipv4network.proto'),
     join_paths(common_proto_dir, 'v1', 'ipv6network.proto'),
     join_paths(common_proto_dir, 'v1', 'bicontiguousnetwork.proto'),
@@ -29,6 +30,7 @@ common_protoc_gen = custom_target(
         'ipv6addr.pb.go',
         'iprange.pb.go',
         'ipnetwork.pb.go',
+        'contiguousnetwork.pb.go',
         'ipv4network.pb.go',
         'ipv6network.pb.go',
         'bicontiguousnetwork.pb.go',
