+++ common/commonpb/v1/ipnetwork.go
@@ -1,160 +1,4 @@
 package commonpb
 
-import (
-	"encoding/json"
-	"fmt"
-	"net/netip"
-
-	"github.com/yanet-platform/xnetip"
-)
-
-// NewContiguousIPNetworkFromContiguous creates a ContiguousIPNetwork from
-// an xnetip CIDR block.
-//
-// The conversion is total: the block is masked by its own invariant, so
-// the message carries the base address in the block's family width (four
-// bytes for IPv4, sixteen for IPv6) and the prefix length verbatim. The
-// inverse of ToContiguous.
-func NewContiguousIPNetworkFromContiguous(net xnetip.Contiguous[xnetip.Network]) *ContiguousIPNetwork {
-	return &ContiguousIPNetwork{
-		Addr:      NewIPAddressFromAddr(net.Network().Addr()),
-		PrefixLen: uint32(net.PrefixLen()),
-	}
-}
-
-// NewContiguousIPNetworkFromPrefix creates a ContiguousIPNetwork from a
-// netip.Prefix value, masking off any host bits.
-//
-// Returns an error if prefix is not valid.
-func NewContiguousIPNetworkFromPrefix(prefix netip.Prefix) (*ContiguousIPNetwork, error) {
-	net, ok := xnetip.ContiguousFromPrefix(prefix)
-	if !ok {
-		return nil, fmt.Errorf("invalid prefix")
-	}
-	return NewContiguousIPNetworkFromContiguous(net), nil
-}
-
-// ParseContiguousIPNetwork parses s as a CIDR prefix, masking off any host
-// bits.
-func ParseContiguousIPNetwork(s string) (*ContiguousIPNetwork, error) {
-	prefix, err := netip.ParsePrefix(s)
-	if err != nil {
-		return nil, fmt.Errorf("failed to parse IP network: %w", err)
-	}
-	return NewContiguousIPNetworkFromPrefix(prefix)
-}
-
-// NetworksFromPrefixes creates ContiguousIPNetwork messages from netip.Prefix
-// values, masking off any host bits.
-//
-// Returns an error if any prefix is not valid.
-func NetworksFromPrefixes(prefixes []netip.Prefix) ([]*ContiguousIPNetwork, error) {
-	networks := make([]*ContiguousIPNetwork, 0, len(prefixes))
-	for idx, prefix := range prefixes {
-		network, err := NewContiguousIPNetworkFromPrefix(prefix)
-		if err != nil {
-			return nil, fmt.Errorf("prefixes[%d]: %w", idx, err)
-		}
-		networks = append(networks, network)
-	}
-
-	return networks, nil
-}
-
-// PrefixesFromNetworks converts ContiguousIPNetwork messages to masked
-// netip.Prefix values.
-//
-// Returns an error if any network is malformed.
-func PrefixesFromNetworks(networks []*ContiguousIPNetwork) ([]netip.Prefix, error) {
-	prefixes := make([]netip.Prefix, 0, len(networks))
-	for idx, network := range networks {
-		prefix, err := network.ToPrefix()
-		if err != nil {
-			return nil, fmt.Errorf("prefixes[%d]: %w", idx, err)
-		}
-		prefixes = append(prefixes, prefix)
-	}
-
-	return prefixes, nil
-}
-
-// ToContiguous converts the ContiguousIPNetwork to an xnetip CIDR block.
-//
-// Returns an error if addr is missing or malformed, or if prefix_len
-// exceeds the address family's bit length; the latter wraps
-// xnetip.ErrCIDROverflow. The returned block is masked, so host bits never
-// leak out even if the message was constructed by hand, and the family
-// follows the address width: a sixteen-byte IPv4-mapped address stays
-// IPv6, consistently with IPAddress. The inverse of
-// NewContiguousIPNetworkFromContiguous.
-func (m *ContiguousIPNetwork) ToContiguous() (xnetip.Contiguous[xnetip.Network], error) {
-	addr, err := m.GetAddr().ToAddr()
-	if err != nil {
-		return xnetip.Contiguous[xnetip.Network]{}, fmt.Errorf("failed to parse network address: %w", err)
-	}
-	net, err := xnetip.ContiguousFromCIDR(addr, int(m.GetPrefixLen()))
-	if err != nil {
-		return xnetip.Contiguous[xnetip.Network]{}, fmt.Errorf("failed to build IP network: %w", err)
-	}
-	return net, nil
-}
-
-// ToPrefix converts the ContiguousIPNetwork back to a netip.Prefix value.
-//
-// Returns an error if addr is malformed or if prefix_len exceeds the
-// address family's bit length, exactly when ToContiguous does. The returned
-// prefix is masked, so host bits never leak out even if the message was
-// constructed by hand.
-func (m *ContiguousIPNetwork) ToPrefix() (netip.Prefix, error) {
-	net, err := m.ToContiguous()
-	if err != nil {
-		return netip.Prefix{}, err
-	}
-	return net.Prefix(), nil
-}
-
-// AsLogValue implements xgrpc.ProtoLogValue for compact gRPC logging.
-func (m *ContiguousIPNetwork) AsLogValue() any {
-	prefix, err := m.ToPrefix()
-	if err != nil {
-		return "invalid"
-	}
-
-	return prefix.String()
-}
-
-// contiguousIPNetworkJSON is the JSON wire shape shared by MarshalJSON and
-// UnmarshalJSON.
-type contiguousIPNetworkJSON struct {
-	Network string `json:"network"`
-}
-
-// MarshalJSON serializes the network as a human-readable CIDR string.
-func (m *ContiguousIPNetwork) MarshalJSON() ([]byte, error) {
-	prefix, err := m.ToPrefix()
-	if err != nil {
-		return nil, err
-	}
-	return json.Marshal(contiguousIPNetworkJSON{Network: prefix.String()})
-}
-
-// UnmarshalJSON accepts the network as a CIDR string under the "network"
-// key.
-func (m *ContiguousIPNetwork) UnmarshalJSON(data []byte) error {
-	var raw contiguousIPNetworkJSON
-	if err := json.Unmarshal(data, &raw); err != nil {
-		return err
-	}
-	if raw.Network == "" {
-		return fmt.Errorf("empty IP network is not allowed")
-	}
-
-	parsed, err := ParseContiguousIPNetwork(raw.Network)
-	if err != nil {
-		return err
-	}
-
-	m.Addr = parsed.Addr
-	m.PrefixLen = parsed.PrefixLen
-	return nil
-}
+// This file is deliberately empty. It is kept as a placeholder for
+// upcoming network messages.
